The last rename of a class, and removal of DB references from JavaScript.
