+++ rifServices/src/main/resources/RIFLogger.properties
@@ -50,7 +50,7 @@ rifGenericLibrary.util.RIFMemoryManager=true
 rifServices.system.RIFServiceStartupProperties=true
 rifServices.system.RIFServiceStartupOptions=true
 
-rifServices.restfulWebServices.pg.PGSQLRIFStudyResultRetrievalWebServiceApplication=true
+rifServices.restfulWebServices.pg.StudyResultRetrievalService=true
 rifServices.restfulWebServices.pg.StudySubmissionService=true
 rifServices.restfulWebServices.ms.MSSQLRIFStudyResultRetrievalWebServiceApplication=true
 rifServices.restfulWebServices.ms.MSSQLRIFStudySubmissionWebServiceApplication=true
