+++ rifServices/src/main/webapp/WEB-INF/web.xml
@@ -15,18 +15,7 @@
 	    <param-value>rifServices.restfulWebServices</param-value>
     </init-param>
 
-	<!-- Trying to get logging in Jersey -->
-	<init-param>
-	   <param-name>com.sun.jersey.spi.container.ContainerRequestFilters</param-name>
-	   <param-value>com.sun.jersey.api.container.filter.LoggingFilter</param-value>
-	</init-param>
-
-	<init-param>
-		<param-name>com.sun.jersey.spi.container.ContainerResponseFilters</param-name>
-		<param-value>com.sun.jersey.api.container.filter.LoggingFilter</param-value>
-	</init-param>
-
-	   <load-on-startup>1</load-on-startup>
+	<load-on-startup>1</load-on-startup>
   </servlet>
  
 </web-app>
