+++ rifWebApplication/src/main/webapp/WEB-INF/index.html
@@ -185,7 +185,6 @@
         <script src="utils/services/rifs-util-maptools.js"></script>
         <script src="backend/services/rifs-back-requests.js"></script>
         <script src="backend/services/rifs-back-interceptor.js"></script>
-        <script src="backend/services/rifs-back-database.js"></script>
         <script src="backend/services/rifs-back-parameters.js"></script>
         <script src="backend/services/rifs-back-urls.js"></script>
